# Benchmark and improve sketch join