--- a/pkg/logql/quantile_over_time_sketch_test.go
+++ b/pkg/logql/quantile_over_time_sketch_test.go
@@ -2,6 +2,7 @@ package logql
 
 import (
 	"errors"
+	"math/rand"
 	"testing"
 
 	"github.com/prometheus/prometheus/model/labels"
@@ -107,3 +108,75 @@ func (e errorStepEvaluator) Error() error {
 }
 
 func (e errorStepEvaluator) Explain(Node) {}
+
+func BenchmarkJoinQuantileSketchVector(b *testing.B) {
+	results := make([]ProbabilisticQuantileVector, 1000)
+	for i := range results {
+		results[i] = make(ProbabilisticQuantileVector, 100)
+		for j := range results[i] {
+			results[i][j] = ProbabilisticQuantileSample{
+				T:      int64(i),
+				F:      newRandomSketch(),
+				Metric: []labels.Label{{Name: "foo", Value: "bar"}},
+			}
+		}
+	}
+
+	ev := &sliceStepEvaluator{
+		slice: results[1:],
+		cur:   1,
+	}
+
+	b.ReportAllocs()
+	b.ResetTimer()
+
+	for n := 0; n < b.N; n++ {
+		// Reset step evaluator
+		ev.cur = 1
+		_, err := JoinQuantileSketchVector(true, results[0], ev)
+		require.NoError(b, err)
+	}
+}
+
+func newRandomSketch() sketch.QuantileSketch {
+	r := rand.New(rand.NewSource(42))
+	s := sketch.NewDDSketch()
+	for i := 0; i < 1000; i++ {
+		s.Add(r.Float64())
+	}
+	return s
+}
+
+type sliceStepEvaluator struct {
+	err   error
+	slice []ProbabilisticQuantileVector
+	cur   int
+}
+
+// Close implements StepEvaluator.
+func (*sliceStepEvaluator) Close() error {
+	return nil
+}
+
+// Error implements StepEvaluator.
+func (ev *sliceStepEvaluator) Error() error {
+	return ev.err
+}
+
+// Explain implements StepEvaluator.
+func (*sliceStepEvaluator) Explain(Node) {
+	return
+}
+
+// Next implements StepEvaluator.
+func (ev *sliceStepEvaluator) Next() (ok bool, ts int64, r StepResult) {
+	if ev.cur >= len(ev.slice) {
+		return false, 0, nil
+	}
+
+	r = ev.slice[ev.cur]
+	ts = ev.slice[ev.cur][0].T
+	ev.cur++
+	ok = ev.cur < len(ev.slice)
+	return
+}