**What this PR does / why we need it**:
We saw that `JoinQuantileSketchVector` was using too much memory. This is a small step to improve the allocated memory.

```
pkg: github.com/grafana/loki/pkg/logql
                            │  main.log   │              pool.log              │
                            │   sec/op    │   sec/op     vs base               │
JoinQuantileSketchVector-10   3.603µ ± 4%   2.631µ ± 0%  -26.99% (p=0.002 n=6)

                            │   main.log   │              pool.log               │
                            │     B/op     │     B/op      vs base               │
JoinQuantileSketchVector-10   8.344Ki ± 0%   2.539Ki ± 0%  -69.57% (p=0.002 n=6)

                            │  main.log  │             pool.log             │
                            │ allocs/op  │ allocs/op   vs base              │
JoinQuantileSketchVector-10   109.0 ± 0%   104.0 ± 0%  -4.59% (p=0.002 n=6)
```

**Checklist**
- [ ] Reviewed the [`CONTRIBUTING.md`](https://github.com/grafana/loki/blob/main/CONTRIBUTING.md) guide (**required**)
- [ ] Documentation added
- [ ] Tests updated
- [ ] `CHANGELOG.md` updated
  - [ ] If the change is worth mentioning in the release notes, add `add-to-release-notes` label
- [ ] Changes that require user attention or interaction to upgrade are documented in `docs/sources/setup/upgrade/_index.md`
- [ ] For Helm chart changes bump the Helm chart version in `production/helm/loki/Chart.yaml` and update `production/helm/loki/CHANGELOG.md` and `production/helm/loki/README.md`. [Example PR](https://github.com/grafana/loki/commit/d10549e3ece02120974929894ee333d07755d213)
- [ ] If the change is deprecating or removing a configuration option, update the `deprecated-config.yaml` and `deleted-config.yaml` files respectively in the `tools/deprecated-config-checker` directory. [Example PR](https://github.com/grafana/loki/pull/10840/commits/0d4416a4b03739583349934b96f272fb4f685d15)