@cstyan 2023-12-20
@jeschkies I pushed another commit to use a pool, improves things further and seems like we're still memory safe.

Here pre-k is without the slice capacity change you made, base is with the slice capacity change, and pool is my latest commit.
```
loki/pkg/logql branch  karsten/benchmark-sketch-join [$?] go v1.21.4 [prod-us-central-0]➜ benchstat -alpha 0.01 pre-k base pool
goos: linux
goarch: amd64
pkg: github.com/grafana/loki/pkg/logql
cpu: AMD Ryzen 9 5950X 16-Core Processor
                            │    pre-k     │                base                 │                pool                 │
                            │    sec/op    │    sec/op     vs base               │    sec/op     vs base               │
JoinQuantileSketchVector-32   5.086µ ± ∞ ¹   3.951µ ± ∞ ¹  -22.32% (p=0.008 n=5)   3.401µ ± ∞ ¹  -33.13% (p=0.008 n=5)
¹ need >= 6 samples for confidence interval at level 0.95

                            │     pre-k     │                 base                 │                 pool                 │
                            │     B/op      │     B/op       vs base               │     B/op       vs base               │
JoinQuantileSketchVector-32   8.344Ki ± ∞ ¹   5.117Ki ± ∞ ¹  -38.67% (p=0.008 n=5)   2.517Ki ± ∞ ¹  -69.84% (p=0.008 n=5)
¹ need >= 6 samples for confidence interval at level 0.95

                            │    pre-k    │               base                │               pool                │
                            │  allocs/op  │  allocs/op   vs base              │  allocs/op   vs base              │
JoinQuantileSketchVector-32   109.0 ± ∞ ¹   103.0 ± ∞ ¹  -5.50% (p=0.008 n=5)   103.0 ± ∞ ¹  -5.50% (p=0.008 n=5)
¹ need >= 6 samples for confidence interval at level 0.95
```