@cstyan 2023-12-21
Benchmark still looks the same:
```
goos: linux
goarch: amd64
pkg: github.com/grafana/loki/pkg/logql
cpu: AMD Ryzen 9 5950X 16-Core Processor
                            │    pre-k     │               k-base                │              safe-pool              │
                            │    sec/op    │    sec/op     vs base               │    sec/op     vs base               │
JoinQuantileSketchVector-32   5.086µ ± ∞ ¹   3.958µ ± ∞ ¹  -22.18% (p=0.008 n=5)   3.405µ ± ∞ ¹  -33.05% (p=0.008 n=5)
¹ need >= 6 samples for confidence interval at level 0.95

                            │     pre-k     │                k-base                │              safe-pool               │
                            │     B/op      │     B/op       vs base               │     B/op       vs base               │
JoinQuantileSketchVector-32   8.344Ki ± ∞ ¹   5.117Ki ± ∞ ¹  -38.67% (p=0.008 n=5)   2.517Ki ± ∞ ¹  -69.84% (p=0.008 n=5)
¹ need >= 6 samples for confidence interval at level 0.95

                            │    pre-k    │              k-base               │             safe-pool             │
                            │  allocs/op  │  allocs/op   vs base              │  allocs/op   vs base              │
JoinQuantileSketchVector-32   109.0 ± ∞ ¹   103.0 ± ∞ ¹  -5.50% (p=0.008 n=5)   103.0 ± ∞ ¹  -5.50% (p=0.008 n=5)
¹ need >= 6 samples for confidence interval at level 0.95
```